feat: Add Q update logic to smartcab/agent.py
Added logic to update the Q learning matrix.  The current code remembers
the previous state, action and reward, so it can calculate and update
Q(s,a) in the next iteration of the simulation.  It was too difficult
to determine the next state before allowing the simulation to calculate
it for me.  The code may have bugs in it at this point...