fix: Fix smartcab/agent.py bugs with current_state list and X/Y dist
Fixed bug in current_state assignment code; had to unpack values from
inputs dictionary before submitting the values to be turned into the
current state.  Removed legacy references to Xdist and Ydist variables
that are not currently being used in this revision of the program.
This file still does not run without bugs...