fix: Fixed find_topics function to generate posterior prob. for test set
When training the model on the X_train set, the vocabulary created
for the LDA model was smaller because had less documents.  When I
tried to used this LDA model on the test set, it failed because the
LDA model had different dimensions.  To resolve this I combined the
train and test set temporarily to get a document term matrix with
the right dimensions, then split them again and only trained with
the document term matrix that represented the training set, but with
the full vocabulary.